Move color_bands function into utils
